track: Piksi v2 working again.
Move fixed point conversion into nap code.
Improve commonality between v2 and v3 targets.
Rename some things in v3 tracking to match their v2 counterparts.

+++ src/track.h
@@ -37,7 +37,7 @@ void track_setup(void);
 
 void tracking_send_state(void);
 
-float propagate_code_phase(float code_phase, float carrier_freq, u32 n_samples);
+double propagate_code_phase(double code_phase, double carrier_freq, u32 n_samples);
 
 /* Update interface */
 void tracking_channels_update(u32 channels_mask);
